# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ouvrier-Express",
        "description": "Application de gestion de relation prestataire-client",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Connexion"
                ],
                "summary": "Authentifier un utilisateur",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nom": {
                                                    "type": "string"
                                                },
                                                "prenom": {
                                                    "type": "string"
                                                },
                                                "tel": {
                                                    "type": "string"
                                                },
                                                "adress": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "login": {
                                                    "type": "string"
                                                },
                                                "role": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur d'authentification"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Inscription"
                ],
                "summary": "Ajouter un utilisateur",
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "string"
                                    },
                                    "adress": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "login": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur ajouté avec succès"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Déconnexion"
                ],
                "summary": "Se Déconnecter",
                "operationId": "79854151e609bd20032aa5aa4d5a35c0",
                "responses": {
                    "200": {
                        "description": "Utilisateur déconnecté avec succès"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Connexion"
                ],
                "summary": "Rafraîchir le token d'authentification",
                "operationId": "7d5da2870d3eba95cdd67878ab20da3a",
                "responses": {
                    "200": {
                        "description": "Token rafraîchi avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur d'authentification"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}